Refactor NodeStack and NodePath (#275)
This PR adds new `NodePath` helper functions and pushes some of the responsibilities from the `NodeStack` to the `NodePath`. The former is designed to be a mutable stack that updates as we traverse the tree, whereas the latter is designed to hold an immutable path of nodes finishing in a specific typed node.