Check missing dependencies before `run` CLI command (#776)
This PR adds a new dependency check before running scripts in the `codama run` CLI command. In other words, it'll check that all external dependencies about to be imported from a script are all installed before running these scripts. When some dependencies are missing, it'll prompt the user to install them.